server/AI_Powered_Resume_Analyzer/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
google-generativeai
python-dotenv
pdfplumber
python-docx
pymongo
djongo
PyJWT
beautifulsoup4
requests
